Extend the python path; rather than replacing it
Some distros (i.e. NixOS) require the special PYTHONPATH, so that
the web extension has access to the python packages it wants (i.e. gi).

Previously, the PYTHONPATH was replaced with the web extension path;
meaning it would crash on NixOS.  This instead prepends the web
extension path to the PYTHONPATH.